+++ ConnectfourCode/ConnectfourCode/ConnectfourCode.csproj
@@ -71,8 +71,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="negamax.cs" />
-    <Compile Include="Position.cs" />
+    <Compile Include="NegaMax.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
